Moved prompting code to separate file
